works

+++ agent_build_refactored/managed_packages/managed_packages_builders.py
@@ -27,6 +27,7 @@ from typing import List, Tuple, Optional, Dict, Type
 
 from agent_build_refactored.tools.runner import Runner, RunnerStep, ArtifactRunnerStep, RunnerMappedPath, EnvironmentRunnerStep, DockerImageSpec,GitHubActionsSettings
 from agent_build_refactored.tools.constants import SOURCE_ROOT, DockerPlatform, Architecture
+from agent_build_refactored.tools.steps_libs.constants import IN_DOCKER
 from agent_build_refactored.tools import check_call_with_log
 from agent_build_refactored.prepare_agent_filesystem import build_linux_fhs_agent_files, add_config
 
@@ -866,16 +867,18 @@ class ManagedPackagesBuilder(Runner):
                 last_repo_agent_libs_package_file=args.last_repo_agent_libs_package_file,
             )
 
-            output_path = SOURCE_ROOT / "build"
+            if not IN_DOCKER:
+                output_path = SOURCE_ROOT / "build"
 
-            os.system(f"ls -l {output_path}")
-            if output_path.exists():
-                shutil.rmtree(output_path)
-            shutil.copytree(
-                builder.packages_output_path,
-                output_path,
-                dirs_exist_ok=True,
-            )
+                os.system(f"ls -l {output_path}")
+                if output_path.exists():
+                    shutil.rmtree(output_path)
+                shutil.copytree(
+                    builder.packages_output_path,
+                    output_path,
+                    dirs_exist_ok=True,
+                )
+
         elif args.command == "publish":
             builder.publish_packages_to_packagecloud(
                 packages_dir_path=pl.Path(args.packages_dir),
